requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

snap-plugin-lib-py>=1.0.10,<2
shutilwhich>=1.1.0
versioneer>=0.18
